requirements.txt
python-dotenv==1.0.0
fastapi==0.104.1
Authlib==1.2.1
requests==2.31.0
httpx==0.25.1
itsdangerous==2.1.2
